Fixed various type conversion/comparison warnings.

+++ src/SFML/Audio/SoundFileReaderWav.cpp
@@ -155,7 +155,7 @@ Uint64 SoundFileReaderWav::read(Int16* samples, Uint64 maxCount)
     assert(m_stream);
 
     Uint64 count = 0;
-    while ((count < maxCount) && (m_stream->tell() < m_dataEnd))
+    while ((count < maxCount) && (static_cast<Uint64>(m_stream->tell()) < m_dataEnd))
     {
         switch (m_bytesPerSample)
         {
